separate some functions from peerdfs
