+++ src/nodes/peermr.hh
@@ -1,0 +1,7 @@
+#include "../messages/idatainfo.hh"
+#include "../messages/igroupinfo.hh"
+#include "../messages/iblockinfo.hh"
+
+    void insert_idata(messages::IDataInfo* idata_info);
+    void insert_igroup(messages::IGroupInfo* igroup_info);
+    void insert_iblock(messages::IBlockInfo* iblock_info);
